requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==5.2.10
django-htmx==1.27.0
gunicorn==23.0.0
packaging==25.0
Pillow==11.2.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
openpyxl==3.1.5
whitenoise==6.11.0
